some changes

+++ tools/scripts/hla.01.main.py
@@ -149,7 +149,7 @@ def main(data_dir, project_dir, software_path, database_path, script_path, ref_f
     else:
         print("Skip multiqc, multiqc_report.html files exist!")
 
-    max_parallel_jobs = os.cpu_count()/2
+    max_parallel_jobs = int(os.cpu_count()/2)
     # Step 2 
     if 'hlahd' in software_list:
         hlahd_result_list = find_files_by_suffix("HLA-HD_Result_final.result.txt", project_dir)
